Combine transformer encoder and decoder into single module
